git-svn-id: file:///tmp/g/svn/mediawikiparser@256 17182e90-d3dd-df11-a988-0025b3ada58c

+++ Visualizer/src/visualizer/display/Layouts.java
@@ -85,9 +85,8 @@ public class Layouts
 	{
 		int tableWidth;
 		int tableHeight;
-		
 		tableWidth = (int)Math.sqrt(v.size());
-		tableHeight = (int)Math.ceil(v.size() / tableWidth);
+		tableHeight = (int)Math.ceil((double)v.size() / (double)tableWidth);
 		
 		// determine how much space is needed to show
 		// the labels (authors' names)
